condition and migration updated

+++ database/migrations/2023_08_16_090854_create_packages_table.php
@@ -16,7 +16,7 @@ class CreatePackagesTable extends Migration
         Schema::create('packages', function (Blueprint $table) {
             $table->id();
             $table->string('package_name');
-            $table->integer('insurance_period');
+            $table->decimal('insurance_period',2,1);
 
             $table->string('coverage');
 //            $table->integer('lowest_amount');
